24 Create a command
maya.cmds.tween() will now open the Tweening dialog.
